+++ CHANGELOG.md
@@ -34,11 +34,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 ### Features
-* (build) [#118](https://github.com/classic-terra/core/pull/118) localnet for Apple Silicon
-
-### Features
+* (build) [#101](https://github.com/classic-terra/core/pull/101) Upgrade test
 * (ante) [#103](https://github.com/classic-terra/core/pull/103) Add burn tax split logic
 * (ante) [#107](https://github.com/classic-terra/core/pull/107) Burn Tax Whitelist
+* (build) [#118](https://github.com/classic-terra/core/pull/118) localnet for Apple Silicon
 
 ### Improvements
 * (build) [#93](https://github.com/classic-terra/core/pull/93) Use golang 1.18 and fix ad-hoc security vulnerabilities